Add files via upload
This Reddit Upvote Bot is a software application designed to automate the process of upvoting posts on the popular social news website Reddit. With this bot, users can increase the visibility and popularity of specific posts and make them more likely to reach the front page of the site. The bot is built using Reddit's API and a programming language such as Python, and can be run from a command-line interface. This bot is intended to be used for educational and demonstration purposes only, and the use of bots on Reddit is against the website's terms of service. It is important to always follow online community guidelines and engage in ethical and fair online practices.